[YAPPIOS-66] 그룹 리팩토링 - 그룹에서 사진 가져올 수 있게 OneToMany 설정 - 테스트를 쉽게 하기위해 유저 정보 파라미터로 전달 - 미션에 필요한 repostiory 추가 - 달성률에 관련된 변수 타입을 Integer로 변경

+++ src/main/java/com/yapp/project/organization/domain/dto/OrgDto.java
@@ -30,7 +30,7 @@ public class OrgDto {
         @ApiModelProperty(value = "그룹제목",example = "명상")
         private String title;
         @ApiModelProperty(value = "달성률",example = "87")
-        private Double rate;
+        private Integer rate;
         @ApiModelProperty(value = "이미지",example = "s3/organization/15")
         private String image;
         @ApiModelProperty(value = "참여자",example = "25")
@@ -63,7 +63,7 @@ public class OrgDto {
         @ApiModelProperty(value = "그룹제목",example = "명상")
         private String title;
         @ApiModelProperty(value = "달성률",example = "87")
-        private Double rate;
+        private Integer rate;
         @ApiModelProperty(value = "이렇게해보세",example = "명상을 처음 접해본다면, 호흡부터 시작해보세요.  \n" +
                 "기도를 통해 들어가 발끝까지 뻗어나가는 호흡을 조용히 느껴보며 머리의 생각을 비워보세요.  \n" +
                 "집중이 힘들다면 조용한 음악이나 자연의 소리를 틀어보는 것도 나쁘지 않아요. \n" +
